remove useless codes
